feat: migrate engine v2 to short index and progressive bucket scan

+++ app/src/main/java/com/expedit/rinha2026/domain/SearchResult.java
@@ -2,5 +2,5 @@ package com.expedit.rinha2026.domain;
 
 public final class SearchResult {
     public int fraudCount;
-    public float worstDistance;
+    public long worstDistance;
 }
